Rework reading commands from Serial; Implement getters for ESP; Implement acknowledgments

+++ src/devices/eeprom_map.cpp
@@ -2,7 +2,7 @@
 
 uint8_t EEMEM  fan_pwm_steps_number_address;
 uint16_t EEMEM fan_pwm_frequency_address;
-uint8_t EEMEM  sunraise_duration_minutes_address;
+uint16_t EEMEM  sunraise_duration_minutes_address;
 uint8_t EEMEM  is_alarm_on_address;
 uint8_t EEMEM  alarm_hours_address;
 uint8_t EEMEM  alarm_minutes_address;
